typescript port of the pagination example
This is a direct port of the pagination example to typescript.  It demonstrates how to extend the default types to reflect the plugins used by useTable.   See `./src/types/react-table-local.d.ts`.  In this case, the usePagination plugin is used, and so those related types are merged with the defaults.

Note that there are some settings in this project that should change as soon as the final types are available, but they might be a hint to anyone using these types in ths short term.

1.  This package uses `patch-package` to remove`node_modules/react-table/index.d.ts`, since tsc was finding it despite my best efforts.
2.  The desired types have been copied into `src/types/react-table.d.ts`

Sadly I couldn't work out how to directly reference the index.d.ts that's in root from the example, tsc, really doesn't want to access files outside it's src root.

There are two bug fixes to the index.d.ts file - if people prefer I can submit them in a separate PR, and then update this to match later.