Steps towards dynamic table typing
This feels like it's a step in the right direction, but it also **doesn't
work** :(

FYI all of these changes are in the examples/pagination-ts folder and
use a local declaration file since the one in the library conflicts with
this.

I took the approach of extending the hook functions defined by the
react-table library and adding additional data to the types so that
typeof could then be used to retrieve info about the other related
extensions provided by the library.

e.g. for `usePagination`, we get:

```ts
  type usePagination<
    D extends object,
    P extends Plugins = TUsePagination
  > = PluginConfig & {
    (hooks: Hooks<D, P>): void
    options: nsUsePagination.UsePaginationOptions<D>
    state: nsUsePagination.UsePaginationState<D>
    instance: nsUsePagination.UsePaginationInstanceProps<D, P>
  }

  export const usePagination: usePagination<D, P>
```

Then, that information can be extracted on a case by case basis to
enhance the resulting types:

```ts
  type UnionToIntersection<U> = (U extends any
    ? (k: U) => void
    : never) extends ((k: infer I) => void)
    ? I
    : never

  type PluginConfig = {
    options?: any
    instance?: any
    state?: any
    cell?: any
    column?: any
    columnInstance?: any
    headerGroup?: any
    row?: any
  }

  type Plugins = PluginConfig[]

  type ArrayPick<T extends Plugins, K extends keyof T[number]> = T[number][K]

  type PluginEnhancements<
    P extends Plugins,
    K extends keyof P[number]
  > = UnionToIntersection<NonNullable<ArrayPick<P, K>>>

  export type TableOptions<
    D extends object,
    P extends Plugins
  > = UseTableOptions<D, P> & PluginEnhancements<P, 'options'>
```

I.e. a single declaration of `TableOptions` can be built up of all of the
options added by the various plugins.

That sounds great, and it works to a point.  The problem that I ran
into, and the point at which it all fell apart, was realizing that many
of the extensions return `Row` or `Column` etc., and those in turn
require knowing about the plugin definitions used.  See the `instance`
reference in `usePagination` for an example.

I **think** that the problem now is the recursive nature of the
declarations.

Anyway, for now I've run out of ideas and feel that I've hit a dead end.
I'd really appreciate some more eyes on this, I'm sure that I've missed
something and perhaps some chunk of this and a slightly different
approach can get things unblocked.