Add useAsyncDebounce to filtering example
When filtering large grids (or the browser runs on a slow computer),
typing in the search box can lack a bit of feedback, as the search box
is only updated with the characters the user types together with the
filtered table.

It took me a small while to figure out how to correctly use
useAsyncDebounce and useState to debounce the onChange event handler, so
I thought it might be useful in the filtering example.