feat(useSortBy): Add option to disable sorting of subrows
Add flag to initial state called disableSubRowSortBy which, if true, will not sort the sub rows in a table. This is to increase performance for long tables where each row has sub rows. Sorting of sub rows is done even if the sub rows are not showing, which can lead to performance issues. The default value of this flag is false, so existing tables will not see any changes in how they work.