docs: Remove extra typing for Column Defs
When using an `accessor` type column helper in a top level column defs array where the array is also typed, you get the type error mentioned in #4382. It was mentioned that the solution here is to not type the array, so this PR removes the typing from the example in the docs.