Fix for page resetting to initialState pageIndex
While using this example to get data from my server, I noticed that when I go to a certain page, it will load the data of that page but then reload back to the pageIndex mentioned in initialState. I noticed that although this fix exists, it hasn't been implemented for this example nor is it mentioned in the docs (https://react-table-v7.tanstack.com/docs/api/useTable).