fix: support unsetting getPaginationRowModel
Previously, it was not possible to unset the table.options.getPaginationRowModel option,
because the result of that function is cached in table._getPaginationRowModel, and therefore the
`!table._getPaginationRowModel` check will always evaluate to false.

Use case: Based on a UI switch, we want to enable or disable the pagination of the React table.

Signed-off-by: Andreas Gerstmayr <agerstmayr@redhat.com>